1-FE: Move stuff out of `src/features` and into `src/shared`
