--- Okane.Client/src/features/navigation/router.constants.ts
@@ -1,5 +1,0 @@
-export enum RouteName {
-  DashboardPage = 'DashboardPage',
-  LoginPage = 'LoginPage',
-  RegisterPage = 'RegisterPage',
-}
