+++ Okane.Client/src/shared/constants/http.constants.ts
